Move utils to utils.js
